sched: packing task to per cpu horse power in periodic balance
Usually the standard pu horse power is scaled as 1024, but some of cpu
are not so powerful, like SMT on Intel cpu. So packing task per logical
 cpu impact performance too much and not power efficient.
If we can pack task per standard cpu power, performance impact will
be relief, and the power efficient is better.

This patch change packing task per logical to per standard cpu power
in peridic balance.

Signed-off-by: Alex Shi <alex.shi@intel.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>